修改描述：1.调整代码结构 2.增加编译环境 3.增加sample代码 影响分析： 是否自测： 自测内容： 测试建议： NOTE：

+++ src/sh_printfmt.cpp
@@ -7,7 +7,7 @@ static void _print_strandblank(char *str)
 {
     int i = 0;
     printf("\033[47;32m%-s\033[0m ",str);
-    for(;i < 32 - strlen(str);i++)
+    for(;i < 32 - (int)strlen(str);i++)
         printf(" ");
 }
